+++ 2_NewDistance/quadtree.h
@@ -262,12 +262,12 @@ private:
         uvec4* nodes_array =  new uvec4[max_node_count_];
         if (settings.poly_type == TRIANGLES) {
             init_node_count_ = mesh_data_->triangle_count;
-            for (int ctr = 0; ctr < init_node_count_; ++ctr) {
+            for (int ctr = 0; ctr < (int)init_node_count_; ++ctr) {
                 nodes_array[ctr] = uvec4(0, 0x1, uint(ctr*3), 0);
             }
         } else if (settings.poly_type == QUADS) {
             init_node_count_ = 2 * mesh_data_->quad_count;
-            for (int ctr = 0; ctr < init_node_count_; ++ctr) {
+            for (int ctr = 0; ctr < (int)init_node_count_; ++ctr) {
                 nodes_array[2*ctr+0] = uvec4(0, 0x1, uint(ctr*4), 0);
                 nodes_array[2*ctr+1] = uvec4(0, 0x1, uint(ctr*4), 1);
             }
